+++ includes/class-mjb-applications.php
@@ -98,6 +98,14 @@ class MJB_Applications
             MJB_Notices::redirect($redirect_url, 'error_missing_fields');
         }
 
+        if (MJB_Application_Guard::is_honeypot_triggered()) {
+            MJB_Notices::redirect($redirect_url, 'error_spam');
+        }
+
+        if (!MJB_Recaptcha::verify()) {
+            MJB_Notices::redirect($redirect_url, 'error_recaptcha');
+        }
+
         if (MJB_Application_Guard::is_rate_limited()) {
             MJB_Notices::redirect($redirect_url, 'error_rate_limited');
         }